python-core/requirements.txt
numpy>=1.24.0
Pillow>=10.0.0
numpy-stl>=3.0.0
scipy>=1.11.0
pyinstaller>=6.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
